requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
psycopg2-binary
Pillow
WTForms==3.1.1
email-validator==2.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
cloudinary==1.36.0
Flask-Login==0.6.3
werkzeug==3.0.1
Flask-Mail==0.9.1